disable rtti and exceptions for msvc (#5167)
* disable rtti and exceptions for msvc

* warnings--

* erff

* arch sse2 for 32bit build

* enable rtti for cross compiling

+++ src/layer/arm/arm_usability.h
@@ -17,7 +17,7 @@
 
 static inline signed char float2int8(float v)
 {
-    int int32 = round(v);
+    int int32 = (int)roundf(v);
     if (int32 > 127) return 127;
     if (int32 < -127) return -127;
     return (signed char)int32;
